--- subsystem/src/test/java/org/wildfly/extension/elytron/tls/SubsystemParsingTestCase.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2023 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.wildfly.extension.elytron.tls;
-
-/**
- * Tests all management expects for subsystem, parsing, marshaling, model definition and other
- * Here is an example that allows you a fine grained controller over what is tested and how. So it can give you ideas what can be done and tested.
- * If you have no need for advanced testing of subsystem you look at {@link ElytronTlsSubsystem1_0TestCase} that tests same stuff but most of the code
- * is hidden inside of test harness
- *
- * @author <a href="kabir.khan@jboss.com">Kabir Khan</a>
- * @author <a href="mailto:carodrig@redhat.com">Cameron Rodriguez</a>
- */
-public class SubsystemParsingTestCase {
-}
